backend/requirements.txt
Flask==3.0.3
# Install ultralytics directly from GitHub for NumPy 2.0 compatibility
git+https://github.com/ultralytics/ultralytics.git
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==2.1.3